requirements.txt
discord.py>=2.4,<3.0
google-api-python-client>=2.175,<3.0
google-auth>=2.35,<3.0
